Add a framework for functional tests.
Make it possible to run functional tests by launching pssht in the background
and then calling OpenSSH or PuTTY.

For now, it is used to test ciphers and message authentication code algorithms.
I plan to rely more on it in the future.

Note: the PuTTY driver lacks a few features compared to the OpenSSH one.
      I plan to address this at a later time. Also, I'm still looking for a
      way to extend this framework to support Windows (with PuTTY).